requirements.txt
# requirements.txt (Updated)
django==5.1.7
gunicorn
whitenoise
sqlparse
tzdata
django-modeltranslation
django-recaptcha
django-widget-tweaks
django-ckeditor==6.7.1
django-compressor